sync-service/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
schedule==1.2.0